Ensure get_full_path resolves correctly (#20)
Prior to this patch, path fragments passed into the get_full_path
function were not actually resolved fully. Instead, they were blindly
concatenated with the path to __file__ if the string was not fully
qualified. However, not all un-fully-qualified paths are relative to the
current directory: for example, "~/foo" would be treated incorrectly by
the pre-patch code.