OD-594 Handle rating as rid edge case (#24)
* Handle rating as rid edge case

Rarely, eFiction will use the ratings themselves as the rid rather than
the number.  If we don't handle this case in conversion, no stories will
successfully be tagged with the rating.

ps-id: 71E667E9-7F7E-4D53-A385-576B3C606D96

* OD-594 Add tests for string ratings

* Explicitly cast function to return a bool

If count is an empty array, it will return that empty array instead of a
boolean. Though [] is falsy, it's a violation of the function signature.

Co-authored-by: Ariana <ariana@astele.co.uk>